Changed ArObjects BottomSheet to Fragment
Changed the arobjects bottomsheet to a fragment. Fixed the bug of the bottomsheet behavior.
