+++ app/src/main/java/com/example/a442projects_thisappslaps_co/MainActivity.java
@@ -4,21 +4,15 @@ import android.Manifest;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.ImageButton;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
-import androidx.recyclerview.widget.GridLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.a442projects_thisappslaps_co.ARObjects.ARObjectsController;
-import com.google.android.material.bottomsheet.BottomSheetBehavior;
+import com.example.a442projects_thisappslaps_co.ARObjects.ARFragment;
 
-import java.util.ArrayList;
 import com.example.a442projects_thisappslaps_co.Shop.ShopFragment;
 import com.example.a442projects_thisappslaps_co.Settings.SettingsFragment;
 import com.example.a442projects_thisappslaps_co.Explore.ExploreFragment;
@@ -32,8 +26,6 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 
     private static int MY_CAMERA_PERMISSIONS;
 
-    private BottomSheetBehavior mBottomSheetBehavior;
-    private RecyclerView mARObjectsRecyclerView;
     private ImageButton mARObjectsImageButton;
     private ImageButton mGalleryImageButton;
     private ImageButton mShopImageButton;
@@ -61,12 +53,9 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 
         initializeViewVariables();
         setListeners();
-        setARObjectsAdapter();
     }
 
     private void initializeViewVariables() {
-        mBottomSheetBehavior = BottomSheetBehavior.from(findViewById(R.id.ar_objects_bottom_sheet));
-        mARObjectsRecyclerView = findViewById(R.id.ar_objects_recycler_view);
         mARObjectsImageButton = findViewById(R.id.ar_objects_image_btn);
         mGalleryImageButton = findViewById(R.id.gallery_image_btn);
         mShopImageButton = findViewById(R.id.shop_image_button);
@@ -80,12 +69,6 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         mShopImageButton.setOnClickListener(this);
         mSettingsImageButton.setOnClickListener(this);
         mExploreImageButton.setOnClickListener(this);
-    }
-
-    private void setARObjectsAdapter() {
-        mARObjectsRecyclerView.setLayoutManager(new GridLayoutManager(this, 3));
-        mARObjectsRecyclerView.setAdapter(
-                new ARObjectsAdapter(new ARObjectsController().createARObjectsDummyList()));
     }
 
     private void requestPermissionForCamera() {
@@ -116,7 +99,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     public void onClick(View view) {
         switch (view.getId()) {
             case R.id.ar_objects_image_btn:
-                mBottomSheetBehavior.setState(BottomSheetBehavior.STATE_EXPANDED);
+                startFragment(new ARFragment(), true);
                 break;
             case R.id.gallery_image_btn:
                 startFragment(new GalleryFragment(), true);
@@ -132,42 +115,6 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                 break;
             default:
                 break;
-        }
-    }
-
-    private class ARObjectsHolder extends RecyclerView.ViewHolder {
-
-        ARObjectsHolder(LayoutInflater inflater, ViewGroup viewGroup) {
-            super(inflater.inflate(R.layout.ar_object_item, viewGroup, false));
-        }
-
-        void bind(Integer drawableRes) {
-            itemView.setBackgroundResource(drawableRes);
-        }
-    }
-
-    private class ARObjectsAdapter extends RecyclerView.Adapter<ARObjectsHolder> {
-
-        private ArrayList<Integer> mARObjectDrawableResList;
-
-        ARObjectsAdapter(ArrayList<Integer> arObjectDrawableResList) {
-            mARObjectDrawableResList = arObjectDrawableResList;
-        }
-
-        @NonNull
-        @Override
-        public ARObjectsHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-            return new ARObjectsHolder(LayoutInflater.from(getApplicationContext()), parent);
-        }
-
-        @Override
-        public void onBindViewHolder(@NonNull ARObjectsHolder holder, int position) {
-            holder.bind(mARObjectDrawableResList.get(position));
-        }
-
-        @Override
-        public int getItemCount() {
-            return mARObjectDrawableResList.size();
         }
     }
