+++ app/src/main/res/layout/activity_main.xml
@@ -43,7 +43,6 @@
         app:layout_constraintBottom_toTopOf="@id/ar_objects_image_btn"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
-
   
     <ImageButton
         android:id="@+id/gallery_image_btn"
@@ -85,32 +84,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         tools:layout_editor_absoluteX="134dp" />
-
-    <androidx.coordinatorlayout.widget.CoordinatorLayout
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/bottom_sheet_top_guideline">
-
-        <LinearLayout
-            android:id="@+id/ar_objects_bottom_sheet"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@color/colorGrey"
-            android:orientation="vertical"
-            app:behavior_peekHeight="0dp"
-            app:layout_behavior="com.google.android.material.bottomsheet.BottomSheetBehavior">
-
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/ar_objects_recycler_view"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:layout_gravity="center_horizontal"
-                android:padding="15dp" />
-
-        </LinearLayout>
-
-    </androidx.coordinatorlayout.widget.CoordinatorLayout>
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/bottom_sheet_top_guideline"
