+++ app/src/main/res/values/strings.xml
@@ -18,4 +18,5 @@
     <string name="user_name">User Name</string>
     <string name="account">My Account</string>
     <string name="explore_text_view">Explore</string>
+    <string name="ar_objects">AR Objects</string>
 </resources>
